Format and mypy

+++ cirq/contrib/routing/device.py
@@ -50,7 +50,7 @@ def _manhattan_distance(qubit1: cirq.GridQubit, qubit2: cirq.GridQubit) -> int:
 
 
 def nx_qubit_layout(graph: nx.Graph) \
-        -> Dict[cirq.Qid, Tuple[int, int]]:
+        -> Dict[cirq.Qid, Tuple[float, float]]:
     """Return a layout for a graph for nodes which are qubits.
 
     This can be used in place of nx.spring_layout or other networkx layouts.
@@ -64,7 +64,7 @@ def nx_qubit_layout(graph: nx.Graph) \
     >>> nx.draw_networkx(g, pos=pos)
 
     """
-    pos: Dict[cirq.Qid, Tuple[int, int]] = {}
+    pos: Dict[cirq.Qid, Tuple[float, float]] = {}
 
     _node_to_i_cache = None
     for node in graph.nodes:
@@ -75,7 +75,9 @@ def nx_qubit_layout(graph: nx.Graph) \
             pos[node] = (node.x, 0.5)
         else:
             if _node_to_i_cache is None:
-                _node_to_i_cache = {n: i for i, n in enumerate(sorted(graph.nodes))}
+                _node_to_i_cache = {
+                    n: i for i, n in enumerate(sorted(graph.nodes))
+                }
             # Position in a line according to sort order
             # Offset to avoid overlap with gridqubits
             pos[node] = (0.5, _node_to_i_cache[node] + 1)
